refactor: reorganize modeling modules into model-specific packages
