+++ examples/llm_inference.py
@@ -11,8 +11,8 @@ from mojo_opset.utils.hf_utils import _resolve_local_files_only
 from mojo_opset.utils.hf_utils import build_model_from_hf
 
 ARCH_MAP = {
-    "Qwen3ForCausalLM": ("mojo_opset.modeling.mojo_qwen3_dense", "Qwen3ForCausalLM"),
-    "SeedOssForCausalLM": ("mojo_opset.modeling.mojo_seed_oss_base", "SeedOssForCausalLM"),
+    "Qwen3ForCausalLM": ("mojo_opset.modeling.qwen3.mojo_qwen3_dense", "Qwen3ForCausalLM"),
+    "SeedOssForCausalLM": ("mojo_opset.modeling.seed_oss.mojo_seed_oss_base", "SeedOssForCausalLM"),
 }
