+++ tests/test_qwen3_dense_patching.py
@@ -2,7 +2,7 @@ from typing import Tuple
 
 import torch
 
-from mojo_opset.modeling import torch_qwen3_dense
+from mojo_opset.modeling.qwen3 import torch_qwen3_dense
 from mojo_opset.utils.logging import get_logger
 
 logger = get_logger(__name__)
